refactor: move test data to improve project structure
